backend/requirements.txt
Django
Pillow
djangorestframework
django-extensions
watchdog
django-watchman
djangorestframework-simplejwt
django-cors-headers
django-environ
psycopg2-binary
environs
channels
channels_redis
drf-spectacular
django-allauth
requests_oauthlib
argon2-cffi
requests
Daphne
asgiref
qrcode
pyotp
olefile
python-dotenv
pygments
tqdm
colorlog
slack_sdk
httpx
pytz
celery
django-celery-beat
redis